+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Version v0.2.1.x (2018-08-30)
+## Functions
+
+### ConvertFrom-Base64
+* Initial Release
+
+### ConvertFrom-MemoryStream
+* Initial Release
+
 # Version v0.2.0.x (2018-07-24)
 ## Functions